fix(browser): keep drawer controls reachable and share tab strip layout
